requirements.txt
# Основные зависимости
python-telegram-bot>=20.4  # Библиотека для работы с Telegram Bot API
aiohttp>=3.8.5             # Асинхронный HTTP клиент для работы с API
python-dotenv>=1.0.0       # Для работы с переменными окружения
pydantic>=2.0.0            # Валидация данных
SQLAlchemy>=2.0.0          # ORM для работы с базой данных
alembic>=1.11.0            # Миграции для базы данных
tiktoken>=0.5.0            # Подсчет токенов для моделей Anthropic (опционально)
anthropic>=0.5.0           # Официальный SDK для Anthropic Claude API

# Обработка файлов
PyPDF2>=3.0.0              # Обработка PDF файлов
python-docx>=0.8.11        # Работа с DOCX файлами
openpyxl>=3.1.0            # Работа с Excel файлами
pandas>=2.0.0              # Обработка табличных данных
Pillow>=10.0.0             # Обработка изображений

# Хранение файлов
boto3>=1.28.0              # Для интеграции с AWS S3 (опционально)

# Логирование и мониторинг
structlog>=23.1.0          # Структурированное логирование
prometheus-client>=0.17.0  # Мониторинг (опционально)

# Для разработки
pytest>=7.4.0              # Тестирование
black>=23.7.0              # Форматирование кода
isort>=5.12.0              # Сортировка импортов
flake8>=6.1.0              # Линтер
pytest-asyncio>=0.21.1     # Поддержка асинхронных тестов
